requirements.txt
rich
pyyaml
flask
